requirements.txt
django>=5.1,<6.0
openai-whisper>=20231117
